refactor: rename same mod name to differnet one
