+++ backend/tauri/src/config/core.rs
