+++ backend/tauri/src/core/mod.rs
@@ -1,5 +1,5 @@
 pub mod clash_api;
-mod core;
+mod clash_core;
 pub mod handle;
 pub mod hotkey;
 pub mod logger;
@@ -11,4 +11,4 @@ pub mod tray;
 pub mod updater;
 pub mod win_service;
 pub mod win_uwp;
-pub use self::core::*;
+pub use self::clash_core::*;
